requirements.txt
esper==3.4
pygame-ce==2.5.3
